requested changes
MangaSee plugin just extends MangaLife
